Refactor utils directory structure
