+++ utils/core/app_state.h
@@ -1,0 +1,2 @@
+#pragma once
+extern bool g_multiRobloxEnabled;
