+++ utils/threading.h
@@ -1,16 +1,2 @@
 #pragma once
-#include <thread>
-#include <utility>
-
-namespace Threading {
-    // Launches f(args...) on a detached background thread.
-    template<typename Func, typename... Args>
-    void newThread(Func &&f, Args &&... args) {
-        std::thread(
-            [fn = std::forward<Func>(f),
-                tup = std::make_tuple(std::forward<Args>(args)...)]() mutable {
-                std::apply(fn, tup);
-            }
-        ).detach();
-    }
-}
+#include "system/threading.h"
